Move to extensions to a different package
