+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/embeddedcomponentmanagerloader/EmbeddedComponentManagerLoaderState.kt
@@ -2,8 +2,8 @@ package com.stripe.android.connect.example.ui.embeddedcomponentmanagerloader
 
 import com.stripe.android.connect.EmbeddedComponentManager
 import com.stripe.android.connect.PrivateBetaConnectSDK
-import com.stripe.android.connect.example.ui.common.Async
-import com.stripe.android.connect.example.ui.common.Uninitialized
+import com.stripe.android.connect.example.core.Async
+import com.stripe.android.connect.example.core.Uninitialized
 
 @OptIn(PrivateBetaConnectSDK::class)
 data class EmbeddedComponentManagerLoaderState(
